+++ services/web/server/src/simcore_service_webserver/catalog/_exceptions.py
