+++ services/web/server/src/simcore_service_webserver/catalog/plugin.py
@@ -1,6 +1,4 @@
-""" Subsystem to communicate with catalog service
-
-"""
+"""Subsystem to communicate with catalog service"""
 
 import logging
 
@@ -8,7 +6,7 @@ from aiohttp import web
 from pint import UnitRegistry
 from servicelib.aiohttp.application_setup import ModuleCategory, app_module_setup
 
-from . import _handlers, _tags_handlers
+from . import _controller_rest, _tags_handlers
 
 _logger = logging.getLogger(__name__)
 
@@ -24,10 +22,10 @@ def setup_catalog(app: web.Application):
     # ensures routes are names that corresponds to function names
     assert all(  # nosec
         route_def.kwargs["name"] == route_def.handler.__name__  # type: ignore[attr-defined] # route_def is a RouteDef not an Abstract
-        for route_def in _handlers.routes
+        for route_def in _controller_rest.routes
     )
 
-    app.add_routes(_handlers.routes)
+    app.add_routes(_controller_rest.routes)
     app.add_routes(_tags_handlers.routes)
 
     # prepares units registry
